fix: fetch 커스텀 함수 토큰 문제로 인해 필요한 부분에만 적용

+++ src/app/_api/bookmark.ts
@@ -1,3 +1,4 @@
+import { authCheck } from "@/utils/function/authCheck";
 import { fetchWithTokenRenewal } from "@/utils/function/fetchWithTokenRenewal";
 import { BookMarkedAddResponse } from "@/utils/types/bookmark/add";
 import { BookMarkedAllCheckResponse } from "@/utils/types/bookmark/allCheck";
@@ -9,10 +10,13 @@ export type getCheckBookmarkResponse = { isBookmarked: boolean };
 
 export const getCheckBookmark = async ({
   auctionId
-}: getCheckBookmarkParams): Promise<getCheckBookmarkResponse> => {
+}: getCheckBookmarkParams): Promise<getCheckBookmarkResponse | null> => {
   const res = await fetchWithTokenRenewal(
     `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/auctions/bookmarks/${auctionId}`
   );
+  if (res === null) {
+    return null;
+  }
   if (!res.ok) {
     throw new Error("서버 에러 발생");
   }
@@ -23,12 +27,20 @@ export const getCheckBookmark = async ({
 export const addBookmark = async (
   auctionId: number
 ): Promise<BookMarkedAddResponse> => {
-  const res = await fetchWithTokenRenewal(
+  const isTokenValid = authCheck();
+
+  if (!isTokenValid) throw new Error("401");
+
+  const res = await fetch(
     `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/auctions/bookmarks/${auctionId}`,
     {
-      method: "POST"
+      method: "POST",
+      headers: {
+        Authorization: `Bearer ${isTokenValid}`
+      }
     }
   );
+
   if (!res.ok) {
     throw new Error(`${res.status}`);
   }
@@ -39,11 +51,18 @@ export const addBookmark = async (
 
 export const deleteBookmark = async (
   auctionId: number
-): Promise<BookMarkedDeleteResponse> => {
-  const res = await fetchWithTokenRenewal(
+): Promise<BookMarkedDeleteResponse | null> => {
+  const isTokenValid = authCheck();
+
+  if (!isTokenValid) throw new Error("401");
+
+  const res = await fetch(
     `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/auctions/bookmarks/${auctionId}`,
     {
-      method: "DELETE"
+      method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${isTokenValid}`
+      }
     }
   );
   if (!res.ok) {
@@ -55,10 +74,13 @@ export const deleteBookmark = async (
 };
 
 export const getCheckBookmarkList =
-  async (): Promise<BookMarkedAllCheckResponse> => {
+  async (): Promise<BookMarkedAllCheckResponse | null> => {
     const res = await fetchWithTokenRenewal(
       `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/auctions/bookmarks`
     );
+    if (res === null) {
+      return null;
+    }
     if (!res.ok) {
       throw new Error("서버 에러 발생");
     }
